Refactors selectors.js -> extract connection/message relevant selectors to their own files
